Rename qp-common to qp-scheduler
