+++ primitives/scheduler/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "qp-common"
-description = "Common primitives for the Quantus blockchain"
+name = "qp-scheduler"
+description = "Common primitives for the scheduler pallet and its dependencies"
 version = "0.1.0"
 license = "Apache-2.0"
 authors.workspace = true
