+++ runtime/Cargo.toml
@@ -58,7 +58,7 @@ pallet-qpow = { workspace = true, default-features = false}
 pallet-wormhole = { workspace = true, default-features = false}
 pallet-reversible-transfers = { workspace = true, default-features = false }
 pallet-vesting = { workspace = true, default-features = false }
-qp-common = { path = "../primitives/common", default-features = false }
+qp-scheduler = { workspace = true, default-features = false }
 sp-consensus-qpow = { workspace = true, default-features = false}
 sp-faucet = {workspace = true, default-features = false}
 log = { workspace = true }
@@ -132,7 +132,7 @@ std = [
 	"dilithium-crypto/full_crypto",
 	"poseidon-resonance/std",
 	"pallet-reversible-transfers/std",
-	"qp-common/std",
+	"qp-scheduler/std",
 ]
 
 runtime-benchmarks = [
